% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texarray.R
\name{texarray}
\alias{texarray}
\title{Convert a matrix to LaTeX table format.}
\usage{
texarray(mat, barpos = 0, bracket = "", small = FALSE)
}
\arguments{
\item{mat}{A matrix to convert to LaTeX format.}

\item{barpos}{The number of columns from the right where to place the vertical line. Defaults to 0 (no vertical line).}

\item{bracket}{The type of bracket to use for the array. Defaults to none.}

\item{small}{Boolean flag indicating whether to use the small matrix format for inline use. Defaults to FALSE.}
}
\value{
A character vector containing the LaTeX code for the matrix in \code{array} environment format.
}
\description{
Converts a matrix to a LaTeX table format using the \code{array} environment.
}
\examples{
mat <- matrix(1:6, nrow = 2)
texarray(mat)

}
\author{
Gabe Osmo
}
